fix task ing22

+++ app/src/main/java/com/nightx/ingale/feature_local/main/data/repository/LocalMainRepositoryImpl.kt
@@ -173,10 +173,12 @@ class LocalMainRepositoryImpl(
                                 setReadable(true, true)
                             }
                             val file = File(appDataDir, previewName)
-                            previewPath = "${appDataDir.path}/$previewName"
-                            val fos = FileOutputStream(file)
-                            albumArt.compress(ImageCompressFormat, 100, fos)
-                            fos.close()
+                            if (isFileValid(file)) continue
+                                previewPath = "${appDataDir.path}/$previewName"
+                                val fos = FileOutputStream(file)
+                                albumArt.compress(ImageCompressFormat, 100, fos)
+                                fos.close()
+
                         } catch (e: IOException) {
                             e.printStackTrace()
                         }
@@ -214,5 +216,15 @@ class LocalMainRepositoryImpl(
                 }
             }
         }
+    }
+
+    private fun isFileValid(file: File): Boolean {
+        // Check if the file exists
+        if (!file.exists()) {
+            return false
+        }
+
+        // Check if the file size is greater than 0 bytes (not empty)
+        return file.length() != 0L
     }
 }
